pyproject.toml
[project]
name = "mlschool"
version = "0.1.0"
description = "Building AI/ML Systems That Don't Suck"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "awscurl>=0.36",
    "boto3>=1.36.25",
    "click>=8.1.8",
    "geventhttpclient>=2.3.3",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "keras==3.11.3",
    "mcp[cli]>=1.3.0",
    "metaflow==2.18.0",
    "metaflow-card-html>=1.0.2",
    "mlflow[extras]==3.3.2",
    "numpy>=1.26.4",
    "pandas>=2.2.3",
    "pylint>=3.3.2",
    "pytest>=8.3.4",
    "ruff>=0.11.6",
    "evidently>=0.7.4,<0.8.0",
    "scikit-learn>=1.6.1,<1.7.0",
    "tensorflow==2.18.1",
    "langchain-community>=0.3.21",
    "faiss-cpu>=1.11.0.post1",
    "langchain-google-genai>=2.1.9",
    "pyyaml>=6.0.2",
    "litellm>=1.76.1",
    "google-adk>=1.13.0",
    "pytest-asyncio>=1.2.0",
]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"
extend-include = ["*.ipynb"]
extend-exclude = ["justfile"]


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["ANN001", "ANN201", "ANN202", "B006", "BLE001", "COM812", "D203", "D213", "E402", "PD901", "PLR0913", "S101", "S311", "T201"]
extend-select = ["E", "I", "F"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"**/tests/*.py" = ["D103"]

[tool.pyright]
venv = ".venv"
typeCheckingMode = "off"
exclude = ["**/justfile"]

[tool.pyright.diagnosticSeverityOverrides]
reportCallIssue = "off"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath="."
addopts="--maxfail=2 -rf -W ignore::DeprecationWarning"
markers = [
    "integration: marks integration tests (deselect with '-m \"not integration\"')"
]
